need to resolve arbiter host

+++ bash/init-rebuild.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+# `./init-rebuild init` to just edit the mongodbversion 4.4.6
+
+if [[ "$1" == "edit" ]];then
+    INITDB=arnobkumarsaha/mongodb-init:hello1
+    kubectl get mongodbversion 4.4.6 -oyaml | sed "s~kubedb/mongodb-init:4.2-v3~"$INITDB"~g" | kubectl replace -f -
+    exit 0
+fi 
 
 docker rmi $(docker images -f dangling=true -q)
